backend/pyproject.toml
[project]
name = ""
version = "0.1.0"
description = ""
authors = [
    {name = "Miroslav Hrabal", email = "mhn@email.cz"},
]
dependencies = [
    "tensorflow",
    "pika",
    "Pillow",
    "lz4",
    "av",
    "Flask",
    "pymongo",
    "click",
    "dramatiq[redis,watch]>=1.13.0",
    "more-itertools>=8.14.0",
    "toolz>=0.12.0",
    "transformers>=4.22.2",
    "torch>=1.12.1",
    "torchaudio>=0.12.1",
    "torchvision>=0.13.1",
    "pydantic>=1.10.2",
    "simple-file-checksum>=1.2.2",
    "attrs>=22.1.0",
    "timm>=0.6.11",
    "datasets>=2.6.1",
    "sortedcontainers>=2.4.0",
    "flask-restx>=1.0.3",
]
requires-python = ">=3.10,<3.11"
license = {text = "MIT"}
readme = "README.md"
[project.optional-dependencies]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[project.urls]

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "flake8>=5.0.4",
    "autopep8>=1.7.0",
    "black>=22.8.0",
    "pytest>=7.1.3",
    "ipython>=8.5.0",
    "isort>=5.10.1",
    "mypy>=0.971",
    "pylint>=2.15.3",
    "pre-commit>=2.20.0",
]
[tool.pdm.scripts]
api = "pdm run python backend/app.py api"
[tool.isort]
profile = "black"
